Add retries to sketch service call
kb_mash is failing due to assembly homology service timeouts, which causes the
sketch service to return an error. It's not clear ATM why the assembly homology
service is timing out, but per Sebastian it's intermittent so add an
expotential backoff to try and allow the app to keep going.

Unfortunately this is a real pain to test, so the plan for now is several
thorough code reviews and test it in CI for a bit, hopefully getting some
retries in the app logs.